refactor: finish query read source extraction
