+++ crates/jazz/src/node/query_eval.rs
@@ -197,9 +197,9 @@ enum CurrentQueryProgramOutput {
     MaintainedView,
 }
 
-mod source_resolution;
+mod read_sources;
 
-use source_resolution::*;
+use read_sources::*;
 
 mod normalization;
 
@@ -3306,735 +3306,6 @@ fn query_order_value(row: &CurrentRow, table: &TableSchema, column: &str) -> Opt
         return row.raw_field(column);
     }
     row.cell(table, column)
-}
-
-fn current_row_fields(table: &TableSchema) -> Vec<String> {
-    let mut fields = vec!["row_uuid".to_owned()];
-    fields.extend(
-        table
-            .columns
-            .iter()
-            .map(|column| user_column_field(&column.name)),
-    );
-    fields.push("$createdBy".to_owned());
-    fields.push("$createdAt".to_owned());
-    fields.push("$updatedBy".to_owned());
-    fields.push("$updatedAt".to_owned());
-    fields.push("tx_time".to_owned());
-    fields.push("tx_node_id".to_owned());
-    fields
-}
-
-fn global_current_storage_fields(
-    table: &TableSchema,
-    include_version: bool,
-    include_settle_position: bool,
-) -> Vec<String> {
-    let mut fields = vec!["row_uuid".to_owned()];
-    if include_version {
-        fields.extend([
-            "schema_version".to_owned(),
-            "parents".to_owned(),
-            "authored_columns".to_owned(),
-        ]);
-    }
-    fields.extend(
-        table
-            .columns
-            .iter()
-            .map(|column| user_column_field(&column.name)),
-    );
-    fields.push("created_by".to_owned());
-    fields.push("created_at".to_owned());
-    fields.push("updated_by".to_owned());
-    fields.push("updated_at".to_owned());
-    fields.push("tx_time".to_owned());
-    fields.push("tx_node_id".to_owned());
-    if include_settle_position {
-        fields.push("global_seq".to_owned());
-    }
-    fields
-}
-
-fn current_row_descriptor(table: &TableSchema) -> RecordDescriptor {
-    RecordDescriptor::new(
-        std::iter::once(("row_uuid".to_owned(), ValueType::Uuid))
-            .chain(table.columns.iter().map(|column| {
-                (
-                    user_column_field(&column.name),
-                    ValueType::Nullable(Box::new(column.column_type.clone())),
-                )
-            }))
-            .chain([
-                ("$createdBy".to_owned(), ValueType::Uuid),
-                ("$createdAt".to_owned(), ValueType::U64),
-                ("$updatedBy".to_owned(), ValueType::Uuid),
-                ("$updatedAt".to_owned(), ValueType::U64),
-                ("tx_time".to_owned(), ValueType::U64),
-                ("tx_node_id".to_owned(), ValueType::U64),
-            ]),
-    )
-}
-
-fn empty_authorized_row_id_graph() -> GraphBuilder {
-    GraphBuilder::inline_records(
-        RecordDescriptor::new([("row_uuid", ValueType::Uuid)]),
-        Vec::<Vec<u8>>::new(),
-    )
-}
-
-fn inline_current_record(
-    table: &TableSchema,
-    descriptor: &RecordDescriptor,
-    row: &CurrentRow,
-) -> Result<Vec<u8>, Error> {
-    let mut values = Vec::with_capacity(table.columns.len() + 7);
-    values.push(Value::Uuid(row.row_uuid().0));
-    for column in &table.columns {
-        values.push(Value::Nullable(row.cell(table, &column.name).map(Box::new)));
-    }
-    if let Some(provenance) = row.provenance()? {
-        values.push(Value::Uuid(provenance.created_by.0));
-        values.push(Value::U64(provenance.created_at.0));
-        values.push(Value::Uuid(provenance.updated_by.0));
-        values.push(Value::U64(provenance.updated_at.0));
-    } else {
-        values.push(Value::Uuid(AuthorId::SYSTEM.0));
-        values.push(Value::U64(0));
-        values.push(Value::Uuid(AuthorId::SYSTEM.0));
-        values.push(Value::U64(0));
-    }
-    let (tx_time, tx_node_alias) = row
-        .projected_tx_alias()
-        .unwrap_or((TxTime(0), NodeAlias(0)));
-    values.push(Value::U64(tx_time.0));
-    values.push(Value::U64(tx_node_alias.0));
-    Ok(descriptor.create(&values)?)
-}
-
-pub(super) fn inline_current_graph(
-    table: &TableSchema,
-    rows: Vec<CurrentRow>,
-) -> Result<GraphBuilder, Error> {
-    let descriptor = current_row_descriptor(table);
-    let records = rows
-        .iter()
-        .map(|row| inline_current_record(table, &descriptor, row))
-        .collect::<Result<Vec<_>, _>>()?;
-    Ok(GraphBuilder::inline_records(descriptor, records))
-}
-
-fn inline_current_graph_with_source_metadata(
-    table: &TableSchema,
-    rows: Vec<CurrentRow>,
-    schema_version_alias: SchemaVersionAlias,
-    coverage: &str,
-    requirements: &SourceRequirements,
-) -> Result<
-    (
-        GraphBuilder,
-        RecordDescriptor,
-        BTreeMap<SourceMetadataRequirement, SourceMetadataFields>,
-    ),
-    Error,
-> {
-    let mut metadata = BTreeMap::new();
-    if requirements
-        .metadata
-        .contains(&SourceMetadataRequirement::VersionWitnesses)
-    {
-        metadata.insert(
-            SourceMetadataRequirement::VersionWitnesses,
-            SourceMetadataFields::VersionWitnesses {
-                schema_version_field: "schema_version".to_owned(),
-                tx_time_field: "tx_time".to_owned(),
-                tx_node_field: "tx_node_id".to_owned(),
-                branch_or_prefix_field: None,
-            },
-        );
-    }
-    if requirements
-        .metadata
-        .contains(&SourceMetadataRequirement::Coverage)
-    {
-        metadata.insert(
-            SourceMetadataRequirement::Coverage,
-            SourceMetadataFields::Coverage {
-                coverage_field: "coverage".to_owned(),
-            },
-        );
-    }
-    if requirements
-        .metadata
-        .contains(&SourceMetadataRequirement::SettlePosition)
-    {
-        metadata.insert(
-            SourceMetadataRequirement::SettlePosition,
-            SourceMetadataFields::SettlePosition {
-                settle_position_field: "settle_position".to_owned(),
-            },
-        );
-    }
-    for requirement in &requirements.metadata {
-        if let SourceMetadataRequirement::Provenance(field) = requirement {
-            metadata.insert(
-                SourceMetadataRequirement::Provenance(*field),
-                SourceMetadataFields::Provenance {
-                    field: source_provenance_field(*field).to_owned(),
-                },
-            );
-        }
-    }
-
-    let descriptor = current_row_descriptor_with_hidden_source_fields(table, &metadata);
-    let records = rows
-        .iter()
-        .map(|row| {
-            inline_current_record_with_source_metadata(
-                table,
-                &descriptor,
-                row,
-                schema_version_alias,
-                coverage,
-            )
-        })
-        .collect::<Result<Vec<_>, _>>()?;
-    Ok((
-        GraphBuilder::inline_records(descriptor.clone(), records),
-        descriptor,
-        metadata,
-    ))
-}
-
-fn inline_current_record_with_source_metadata(
-    table: &TableSchema,
-    descriptor: &RecordDescriptor,
-    row: &CurrentRow,
-    schema_version_alias: SchemaVersionAlias,
-    coverage: &str,
-) -> Result<Vec<u8>, Error> {
-    let mut values = Vec::new();
-    values.push(Value::Uuid(row.row_uuid().0));
-    for column in &table.columns {
-        values.push(Value::Nullable(row.cell(table, &column.name).map(Box::new)));
-    }
-    let provenance = row.provenance()?.unwrap_or(RowProvenance {
-        created_by: AuthorId::SYSTEM,
-        created_at: TxTime(0),
-        updated_by: AuthorId::SYSTEM,
-        updated_at: TxTime(0),
-    });
-    values.extend([
-        Value::Uuid(provenance.created_by.0),
-        Value::U64(provenance.created_at.0),
-        Value::Uuid(provenance.updated_by.0),
-        Value::U64(provenance.updated_at.0),
-    ]);
-    let (tx_time, tx_node_alias) = row
-        .projected_tx_alias()
-        .unwrap_or((TxTime(0), NodeAlias(0)));
-    values.extend([Value::U64(tx_time.0), Value::U64(tx_node_alias.0)]);
-    if descriptor.field_index("table").is_some() {
-        values.extend([
-            Value::String(table.name.clone()),
-            Value::String("content".to_owned()),
-            Value::U64(schema_version_alias.0),
-            Value::Array(Vec::new()),
-            Value::Uuid(provenance.created_by.0),
-            Value::U64(provenance.created_at.0),
-            Value::Uuid(provenance.updated_by.0),
-            Value::U64(provenance.updated_at.0),
-            Value::Nullable(None),
-        ]);
-    }
-    if descriptor.field_index("coverage").is_some() {
-        values.push(Value::String(coverage.to_owned()));
-    }
-    if descriptor.field_index("settle_position").is_some() {
-        values.push(Value::Nullable(None));
-    }
-    Ok(descriptor.create(&values)?)
-}
-
-fn inline_include_deleted_current_graph(
-    table: &TableSchema,
-    rows: Vec<(CurrentRow, bool)>,
-) -> Result<GraphBuilder, Error> {
-    let descriptor = include_deleted_current_row_descriptor(table);
-    let mut records = Vec::with_capacity(rows.len());
-    for (row, deleted) in rows {
-        let mut values = Vec::with_capacity(table.columns.len() + 8);
-        values.push(Value::Uuid(row.row_uuid().0));
-        for column in &table.columns {
-            values.push(Value::Nullable(row.cell(table, &column.name).map(Box::new)));
-        }
-        if let Some(provenance) = row.provenance()? {
-            values.push(Value::Uuid(provenance.created_by.0));
-            values.push(Value::U64(provenance.created_at.0));
-            values.push(Value::Uuid(provenance.updated_by.0));
-            values.push(Value::U64(provenance.updated_at.0));
-        } else {
-            values.push(Value::Uuid(AuthorId::SYSTEM.0));
-            values.push(Value::U64(0));
-            values.push(Value::Uuid(AuthorId::SYSTEM.0));
-            values.push(Value::U64(0));
-        }
-        let (tx_time, tx_node_alias) = row
-            .projected_tx_alias()
-            .unwrap_or((TxTime(0), NodeAlias(0)));
-        values.push(Value::U64(tx_time.0));
-        values.push(Value::U64(tx_node_alias.0));
-        values.push(Value::Bool(deleted));
-        records.push(descriptor.create(&values)?);
-    }
-    Ok(GraphBuilder::inline_records(descriptor, records))
-}
-
-fn inline_branch_current_graph(
-    table: &TableSchema,
-    rows: Vec<(CurrentRow, TxTime, NodeAlias, Option<BranchId>)>,
-    schema_version_alias: SchemaVersionAlias,
-    requirements: &SourceRequirements,
-) -> Result<
-    (
-        GraphBuilder,
-        RecordDescriptor,
-        BTreeMap<SourceMetadataRequirement, SourceMetadataFields>,
-    ),
-    Error,
-> {
-    let mut metadata = BTreeMap::new();
-    if requirements
-        .metadata
-        .contains(&SourceMetadataRequirement::VersionWitnesses)
-    {
-        metadata.insert(
-            SourceMetadataRequirement::VersionWitnesses,
-            SourceMetadataFields::VersionWitnesses {
-                schema_version_field: "schema_version".to_owned(),
-                tx_time_field: "tx_time".to_owned(),
-                tx_node_field: "tx_node_id".to_owned(),
-                branch_or_prefix_field: Some("branch_id".to_owned()),
-            },
-        );
-    }
-    if requirements
-        .metadata
-        .contains(&SourceMetadataRequirement::Coverage)
-    {
-        metadata.insert(
-            SourceMetadataRequirement::Coverage,
-            SourceMetadataFields::Coverage {
-                coverage_field: "coverage".to_owned(),
-            },
-        );
-    }
-    if requirements
-        .metadata
-        .contains(&SourceMetadataRequirement::SettlePosition)
-    {
-        metadata.insert(
-            SourceMetadataRequirement::SettlePosition,
-            SourceMetadataFields::SettlePosition {
-                settle_position_field: "settle_position".to_owned(),
-            },
-        );
-    }
-    for requirement in &requirements.metadata {
-        if let SourceMetadataRequirement::Provenance(field) = requirement {
-            metadata.insert(
-                SourceMetadataRequirement::Provenance(*field),
-                SourceMetadataFields::Provenance {
-                    field: source_provenance_field(*field).to_owned(),
-                },
-            );
-        }
-    }
-    let descriptor = current_row_descriptor_with_hidden_source_fields(table, &metadata);
-    let records = rows
-        .iter()
-        .map(|(row, tx_time, tx_node, branch)| {
-            inline_branch_current_record(
-                table,
-                &descriptor,
-                row,
-                schema_version_alias,
-                (*tx_time, *tx_node),
-                *branch,
-            )
-        })
-        .collect::<Result<Vec<_>, _>>()?;
-    Ok((
-        GraphBuilder::inline_records(descriptor.clone(), records),
-        descriptor,
-        metadata,
-    ))
-}
-
-fn inline_branch_current_record(
-    table: &TableSchema,
-    descriptor: &RecordDescriptor,
-    row: &CurrentRow,
-    schema_version_alias: SchemaVersionAlias,
-    witness: (TxTime, NodeAlias),
-    branch_id: Option<BranchId>,
-) -> Result<Vec<u8>, Error> {
-    let mut values = Vec::new();
-    values.push(Value::Uuid(row.row_uuid().0));
-    for column in &table.columns {
-        values.push(Value::Nullable(row.cell(table, &column.name).map(Box::new)));
-    }
-    let provenance = row.provenance()?.unwrap_or(RowProvenance {
-        created_by: AuthorId::SYSTEM,
-        created_at: TxTime(0),
-        updated_by: AuthorId::SYSTEM,
-        updated_at: TxTime(0),
-    });
-    values.extend([
-        Value::Uuid(provenance.created_by.0),
-        Value::U64(provenance.created_at.0),
-        Value::Uuid(provenance.updated_by.0),
-        Value::U64(provenance.updated_at.0),
-    ]);
-    let (tx_time, tx_node_alias) = witness;
-    values.extend([Value::U64(tx_time.0), Value::U64(tx_node_alias.0)]);
-    if descriptor.field_index("table").is_some() {
-        values.extend([
-            Value::String(table.name.clone()),
-            Value::String("content".to_owned()),
-            Value::U64(schema_version_alias.0),
-            Value::Array(Vec::new()),
-            Value::Uuid(provenance.created_by.0),
-            Value::U64(provenance.created_at.0),
-            Value::Uuid(provenance.updated_by.0),
-            Value::U64(provenance.updated_at.0),
-            Value::Nullable(None),
-        ]);
-        if descriptor.field_index("branch_id").is_some() {
-            values.push(Value::Nullable(
-                branch_id.map(|branch| Box::new(Value::Uuid(branch.0))),
-            ));
-        }
-    }
-    if descriptor.field_index("coverage").is_some() {
-        values.push(Value::String("branch-current".to_owned()));
-    }
-    if descriptor.field_index("settle_position").is_some() {
-        values.push(Value::Nullable(None));
-    }
-    Ok(descriptor.create(&values)?)
-}
-
-#[cfg(test)]
-fn historical_current_graph_full_scan(
-    table: &TableSchema,
-    table_id: PhysicalTableId,
-    position: GlobalSeq,
-    history_rows: GraphBuilder,
-) -> GraphBuilder {
-    let cut_predicate = PredicateExpr::And(vec![
-        PredicateExpr::eq("physical_table_id", Value::U64(table_id.0)),
-        PredicateExpr::LtEq {
-            field: "global_seq".to_owned(),
-            value: Value::U64(position.0).into(),
-        },
-    ])
-    .canonicalize();
-    let changes_for_layer = |layer: &'static str| {
-        GraphBuilder::table("jazz_global_changes").filter(
-            PredicateExpr::And(vec![
-                cut_predicate.clone(),
-                PredicateExpr::eq("layer", Value::Bytes(layer.as_bytes().to_vec())),
-            ])
-            .canonicalize(),
-        )
-    };
-    let nullable_deletion_type = ValueType::Nullable(Box::new(ValueType::EnumTag(
-        groove::records::ScalarEnumSchema::new("jazz_deletion", ["deleted", "restored"])
-            .expect("valid deletion enum")
-            .with_system_registry(groove::records::SystemVariantRegistry::deletion_state()),
-    )));
-    let content_events = changes_for_layer("content").project_fields([
-        ProjectField::named("row_uuid"),
-        ProjectField::named("tx_time"),
-        ProjectField::named("tx_node_id"),
-        ProjectField::literal("event_layer", Value::String("content".to_owned())),
-        ProjectField::null_typed("deletion", nullable_deletion_type.clone()),
-    ]);
-    let register_events = changes_for_layer("deletion").project_fields([
-        ProjectField::named("row_uuid"),
-        ProjectField::named("tx_time"),
-        ProjectField::named("tx_node_id"),
-        ProjectField::literal("event_layer", Value::String("deletion".to_owned())),
-        ProjectField::renamed("_deletion", "deletion"),
-    ]);
-    let latest_event = GraphBuilder::arg_max_by(
-        GraphBuilder::union([content_events.clone(), register_events]),
-        ["row_uuid"],
-        ["tx_time", "tx_node_id"],
-    );
-    let content_winners =
-        GraphBuilder::arg_max_by(content_events, ["row_uuid"], ["tx_time", "tx_node_id"]);
-    let history_rows = history_rows.project(maintained_view_history_storage_field_names(table));
-    let content_current = GraphBuilder::join(
-        history_rows,
-        content_winners,
-        ["row_uuid", "tx_time", "tx_node_id"],
-        ["row_uuid", "tx_time", "tx_node_id"],
-    )
-    .project_fields(
-        ["row_uuid".to_owned()]
-            .into_iter()
-            .chain(
-                table
-                    .columns
-                    .iter()
-                    .map(|column| user_column_field(&column.name)),
-            )
-            .map(|field| ProjectField::renamed(left_field(&field), field))
-            .chain([
-                ProjectField::renamed("left.created_by", "$createdBy"),
-                ProjectField::renamed("left.created_at", "$createdAt"),
-                ProjectField::renamed("left.updated_by", "$updatedBy"),
-                ProjectField::renamed("left.updated_at", "$updatedAt"),
-                ProjectField::renamed("left.tx_time", "tx_time"),
-                ProjectField::renamed("left.tx_node_id", "tx_node_id"),
-            ]),
-    );
-    let latest_content = latest_event.clone().filter(PredicateExpr::eq(
-        "event_layer",
-        Value::String("content".to_owned()),
-    ));
-    let content_is_latest = GraphBuilder::join(
-        content_current.clone(),
-        latest_content,
-        ["row_uuid", "tx_time", "tx_node_id"],
-        ["row_uuid", "tx_time", "tx_node_id"],
-    )
-    .project_fields(
-        current_row_fields(table)
-            .into_iter()
-            .map(|field| ProjectField::renamed(left_field(&field), field)),
-    );
-    let latest_restore = latest_event.filter(
-        PredicateExpr::And(vec![
-            PredicateExpr::eq("event_layer", Value::String("deletion".to_owned())),
-            PredicateExpr::eq(
-                "deletion",
-                Value::Nullable(Some(Box::new(Value::EnumTag(1)))),
-            ),
-        ])
-        .canonicalize(),
-    );
-    let restored_content =
-        GraphBuilder::join(content_current, latest_restore, ["row_uuid"], ["row_uuid"])
-            .project_fields(
-                current_row_fields(table)
-                    .into_iter()
-                    .map(|field| ProjectField::renamed(left_field(&field), field)),
-            );
-    GraphBuilder::union([content_is_latest, restored_content])
-}
-
-fn include_deleted_current_row_descriptor(table: &TableSchema) -> RecordDescriptor {
-    RecordDescriptor::new(
-        std::iter::once(("row_uuid".to_owned(), ValueType::Uuid))
-            .chain(table.columns.iter().map(|column| {
-                (
-                    user_column_field(&column.name),
-                    ValueType::Nullable(Box::new(column.column_type.clone())),
-                )
-            }))
-            .chain([
-                ("$createdBy".to_owned(), ValueType::Uuid),
-                ("$createdAt".to_owned(), ValueType::U64),
-                ("$updatedBy".to_owned(), ValueType::Uuid),
-                ("$updatedAt".to_owned(), ValueType::U64),
-                ("tx_time".to_owned(), ValueType::U64),
-                ("tx_node_id".to_owned(), ValueType::U64),
-            ])
-            .chain([("__jazz_deleted".to_owned(), ValueType::Bool)]),
-    )
-}
-
-fn include_deleted_current_graph(table: &TableSchema, tier: DurabilityTier) -> GraphBuilder {
-    let user_fields = table
-        .columns
-        .iter()
-        .map(|column| user_column_field(&column.name))
-        .collect::<Vec<_>>();
-    let mut content_storage_fields = vec![
-        "row_uuid".to_owned(),
-        "schema_version".to_owned(),
-        "parents".to_owned(),
-        "authored_columns".to_owned(),
-    ];
-    content_storage_fields.extend(user_fields.iter().cloned());
-    content_storage_fields.push("created_by".to_owned());
-    content_storage_fields.push("created_at".to_owned());
-    content_storage_fields.push("updated_by".to_owned());
-    content_storage_fields.push("updated_at".to_owned());
-    content_storage_fields.push("tx_time".to_owned());
-    content_storage_fields.push("tx_node_id".to_owned());
-    let normalize_content_fields = |graph: GraphBuilder| {
-        graph.project_fields(
-            ["row_uuid".to_owned()]
-                .into_iter()
-                .chain(user_fields.iter().cloned())
-                .map(ProjectField::named)
-                .chain([
-                    ProjectField::renamed("created_by", "$createdBy"),
-                    ProjectField::renamed("created_at", "$createdAt"),
-                    ProjectField::renamed("updated_by", "$updatedBy"),
-                    ProjectField::renamed("updated_at", "$updatedAt"),
-                    ProjectField::named("tx_time"),
-                    ProjectField::named("tx_node_id"),
-                ]),
-        )
-    };
-    let edge_visible_ahead = |table_name: String, fields: Vec<String>| {
-        GraphBuilder::join(
-            GraphBuilder::table(table_name).project(fields.clone()),
-            GraphBuilder::table("jazz_transactions")
-                .filter(
-                    PredicateExpr::Or(vec![
-                        PredicateExpr::eq("durability", Value::EnumTag(2)),
-                        PredicateExpr::eq("durability", Value::EnumTag(3)),
-                    ])
-                    .canonicalize(),
-                )
-                .project(["time", "node_id"]),
-            ["tx_time", "tx_node_id"],
-            ["time", "node_id"],
-        )
-        .project_fields(
-            fields
-                .into_iter()
-                .map(|field| ProjectField::renamed(left_field(&field), field)),
-        )
-    };
-    let (content_current, deletion_current) = if tier == DurabilityTier::Global {
-        (
-            normalize_content_fields(
-                GraphBuilder::table(global_current_table_name(&table.name))
-                    .project(content_storage_fields.clone()),
-            ),
-            GraphBuilder::table(register_global_current_table_name(&table.name)),
-        )
-    } else {
-        let ahead_content = if tier == DurabilityTier::Edge {
-            normalize_content_fields(edge_visible_ahead(
-                ahead_current_table_name(&table.name),
-                content_storage_fields.clone(),
-            ))
-        } else {
-            normalize_content_fields(
-                GraphBuilder::table(ahead_current_table_name(&table.name))
-                    .project(content_storage_fields.clone()),
-            )
-        };
-        let deletion_fields = vec![
-            "row_uuid".to_owned(),
-            "tx_time".to_owned(),
-            "tx_node_id".to_owned(),
-            "created_by".to_owned(),
-            "created_at".to_owned(),
-            "updated_by".to_owned(),
-            "updated_at".to_owned(),
-            "_deletion".to_owned(),
-        ];
-        let ahead_deletion = if tier == DurabilityTier::Edge {
-            edge_visible_ahead(
-                register_ahead_current_table_name(&table.name),
-                deletion_fields.clone(),
-            )
-        } else {
-            GraphBuilder::table(register_ahead_current_table_name(&table.name))
-                .project(deletion_fields.clone())
-        };
-        (
-            GraphBuilder::arg_max_by(
-                GraphBuilder::union([
-                    normalize_content_fields(
-                        GraphBuilder::table(global_current_table_name(&table.name))
-                            .project(content_storage_fields.clone()),
-                    ),
-                    ahead_content,
-                ]),
-                ["row_uuid"],
-                ["tx_time", "tx_node_id"],
-            )
-            .project(current_row_fields(table)),
-            GraphBuilder::arg_max_by(
-                GraphBuilder::union([
-                    GraphBuilder::table(register_global_current_table_name(&table.name))
-                        .project(deletion_fields),
-                    ahead_deletion,
-                ]),
-                ["row_uuid"],
-                ["tx_time", "tx_node_id"],
-            ),
-        )
-    };
-    let deleted_winners = deletion_current
-        .filter(PredicateExpr::eq("_deletion", Value::EnumTag(0)))
-        .project_fields([
-            ProjectField::named("row_uuid"),
-            ProjectField::named("tx_time"),
-            ProjectField::named("tx_node_id"),
-            ProjectField::renamed("updated_by", "$updatedBy"),
-            ProjectField::renamed("updated_at", "$updatedAt"),
-        ]);
-    let undeleted = GraphBuilder::anti_join(
-        content_current.clone(),
-        deleted_winners.clone(),
-        ["row_uuid"],
-        ["row_uuid"],
-    )
-    .project_fields(
-        current_row_fields(table)
-            .into_iter()
-            .map(ProjectField::named)
-            .chain([ProjectField::literal("__jazz_deleted", Value::Bool(false))]),
-    );
-    let deleted = GraphBuilder::join(content_current, deleted_winners, ["row_uuid"], ["row_uuid"])
-        .project_fields(
-            current_row_fields(table)
-                .into_iter()
-                .map(|field| {
-                    let source = match field.as_str() {
-                        "$updatedBy" | "$updatedAt" | "tx_time" | "tx_node_id" => {
-                            right_field(&field)
-                        }
-                        _ => left_field(&field),
-                    };
-                    ProjectField::renamed(source, field)
-                })
-                .chain([ProjectField::literal("__jazz_deleted", Value::Bool(true))]),
-        );
-    GraphBuilder::union([undeleted, deleted])
-}
-
-fn maintained_view_history_storage_field_names(table: &TableSchema) -> Vec<String> {
-    let mut fields = vec![
-        "row_uuid".to_owned(),
-        "tx_time".to_owned(),
-        "tx_node_id".to_owned(),
-        "schema_version".to_owned(),
-        "parents".to_owned(),
-        "created_by".to_owned(),
-        "created_at".to_owned(),
-        "updated_by".to_owned(),
-        "updated_at".to_owned(),
-    ];
-    fields.extend(
-        table
-            .columns
-            .iter()
-            .map(|column| user_column_field(&column.name)),
-    );
-    fields.push("authored_columns".to_owned());
-    fields
 }
 
 #[cfg(test)]
